[CCUBE-1058][KT] Refactor typings and helper
