--- src/components/element-editor/types.ts
@@ -1,5 +1,0 @@
-import { IBaseFieldAttributes } from "src/schemas";
-
-export interface IFieldValues extends Omit<IBaseFieldAttributes, "internalId"> {
-    placeholder?: string;
-}
