+++ src/context-providers/builder/index.ts
@@ -1,3 +1,4 @@
 export * from "./hook";
 export * from "./provider";
+export * from "./element.types";
 export * from "./types";
