+++ src/context-providers/builder/types.ts
@@ -1,5 +1,5 @@
 import { Dispatch } from "react";
-import { TElement } from "src/schemas";
+import { TElement } from "./element.types";
 
 export enum EBuilderMode {
     /** Mode where we get to select the elements to add */
