+++ src/schemas/text-based-fields.ts
@@ -1,0 +1,3 @@
+import * as yup from "yup";
+
+export const EMAIL_SCHEMA = yup.object().shape({});
